requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
imbalanced-learn>=0.10.0

# Visualization Libraries
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.25.0
gunicorn>=20.1.0

# Model Interpretability
shap>=0.42.0

# Core Utilities
joblib>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0
requests>=2.28.0
urllib3>=1.26.0

# Security and Cryptography
cryptography>=3.4.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging and Monitoring
loguru>=0.7.0
structlog>=22.1.0
sentry-sdk>=1.9.0
prometheus-client>=0.14.0

# Configuration Management
configparser>=5.2.0

# Process Management
supervisor>=4.2.0
psutil>=5.9.0

# Redis Support
redis>=4.5.0
hiredis>=2.2.0

# Report Generation
reportlab>=4.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
jinja2>=3.1.0
schedule>=1.2.0

# Database Support
sqlalchemy>=1.4.0
alembic>=1.8.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0

# Dashboard-specific requirements
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6

# Deployment and Container Support
docker>=6.0.0

# Container Orchestration (optional)
kubernetes>=24.2.0

# Monitoring and Observability (optional)
grafana-api>=1.0.3

# Cloud Providers (optional - install as needed)
# boto3>=1.24.0  # AWS
# azure-storage-blob>=12.12.0  # Azure
# google-cloud-storage>=2.5.0  # GCP

# CI/CD Integration (optional - install as needed)
# jenkins-job-builder>=3.11.0
# gitlab-python>=3.6.0
# github3.py>=3.2.0